FIX BUG WITH CONNECT GAME  SOCKET:
