+++ backend/app/src/game-socket/Objects/GameServe.ts
@@ -1,11 +1,12 @@
+/* eslint-disable curly */
 /* eslint-disable max-lines-per-function */
 /* eslint-disable max-statements */
 import Player from "./Player";
 import Board from "./Board";
 import Ball from "./Ball";
 import Net from "./Net";
-import { NodeAnimationFrame } from "../GameSocketEvents";
 import {v4 as uuidv4 } from "uuid";
+import { NodeAnimationFrame } from "../NodeAnimationFrame";
 
 class GameServe
 {
@@ -23,7 +24,7 @@ class GameServe
 	public startDisplayed: boolean;
 	public continueAnimating: boolean;
 	public loop: NodeAnimationFrame | undefined;
-	public	countUser: number;
+	public userConnected: number;
 	public displayStartMessage: () => void;
 	public displayEndMessage: () => void;
 	public initPlayers: () => void;
@@ -32,6 +33,7 @@ class GameServe
 
 	public getSeralizable: () => any;
 
+	// public	isGameInstanceEmpty: () => boolean;
 	public constructor(roomName: string)
 	{
 		this.uuid = uuidv4();
@@ -48,7 +50,7 @@ class GameServe
 		this.startDisplayed = true;
 		this.continueAnimating = true;
 		this.loop = undefined;
-		this.countUser = 0;
+		this.userConnected = 0;
 
 		this.initPlayers = () =>
 		{
@@ -73,22 +75,22 @@ class GameServe
 
 		this.getSeralizable = () =>
 		{
-			console.log("getSeralizable inside  GameServe.ts", {
-				uuid: this.uuid,
-				roomName: this.roomName,
-				frameRate: this.frameRate,
-				frameCount: this.frameCount,
-				playerOne: this.playerOne.getSerializable(),
-				plpayerTwo: this.playerTwo.getSerializable(),
-				board: this.board.getSeralizable(),
-				ball: this.ball.getSeralizable(),
-				net: this.net.getSerializable(),
-				scoreLimit: this.scoreLimit,
-				actionKeyPress: this.actionKeyPress,
-				startDisplayed: this.startDisplayed,
-				continueAnimating: this.continueAnimating,
-				loop: this.loop?.getSerializable(),
-			});
+			// console.log("getSeralizable inside  GameServe.ts", {
+			// 	uuid: this.uuid,
+			// 	roomName: this.roomName,
+			// 	frameRate: this.frameRate,
+			// 	frameCount: this.frameCount,
+			// 	playerOne: this.playerOne.getSerializable(),
+			// 	plpayerTwo: this.playerTwo.getSerializable(),
+			// 	board: this.board.getSeralizable(),
+			// 	ball: this.ball.getSeralizable(),
+			// 	net: this.net.getSerializable(),
+			// 	scoreLimit: this.scoreLimit,
+			// 	actionKeyPress: this.actionKeyPress,
+			// 	startDisplayed: this.startDisplayed,
+			// 	continueAnimating: this.continueAnimating,
+			// 	loop: this.loop?.getSerializable(),
+			// });
 			return ({
 				uuid: this.uuid,
 				roomName: this.roomName,
@@ -104,7 +106,7 @@ class GameServe
 				startDisplayed: this.startDisplayed,
 				continueAnimating: this.continueAnimating,
 				loop: this.loop?.getSerializable(),
-				countUser: this.countUser
+				userConnected: this.userConnected
 			});
 		};
 		this.isSocketIdExistGameInstance = (clientId: string) =>
@@ -116,6 +118,27 @@ class GameServe
 				return (true);
 			return (false);
 		};
+		// this.isGameInstanceEmpty = () =>
+		// {
+		// 	const	sockOne = this.playerOne.socketId;
+		// 	const	sockTwo = this.playerTwo.socketId;
+
+		// 	if (sockOne !== "undefined")
+		// 	{
+		// 		if (sockOne !== "disconnected")
+		// 		{
+		// 			return (false);
+		// 		}
+		// 	}
+		// 	if (sockTwo !== "undefined")
+		// 	{
+		// 		if (sockTwo !== "disconnected")
+		// 		{
+		// 			return (false);
+		// 		}
+		// 	}
+		// 	return (true);
+		// };
 	}
 }
